# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API HackR",
        "description": "Voici la documentation de l'API HackR. Vous pouvez tester les différentes routes (Authentification, hack, logs & rules) en utilisant l'interface Swagger.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/getLastLogs": {
            "post": {
                "tags": [
                    "Logs"
                ],
                "summary": "Récupère les derniers logs d'activité",
                "description": "Cette méthode permet à un administrateur de récupérer les derniers logs d'activité, triés par date. Elle nécessite que l'utilisateur soit authentifié et possède le rôle d'administrateur.",
                "operationId": "getLastLogs",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nbLogs"
                                ],
                                "properties": {
                                    "nbLogs": {
                                        "description": "Le nombre de derniers logs à récupérer",
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Les derniers logs récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "user": {
                                                "description": "Nom de l'utilisateur ayant effectué l'action",
                                                "type": "string"
                                            },
                                            "fonctionnalite": {
                                                "description": "Nom de la fonctionnalité associée à l'action",
                                                "type": "string"
                                            },
                                            "date_action": {
                                                "description": "Date et heure de l'action",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "description_action": {
                                                "description": "Description de l'action réalisée",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Utilisateur non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit pour l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vous n'avez pas les droits pour accéder à cette ressource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Message d'erreur détaillé",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/getLogsByUser": {
            "post": {
                "tags": [
                    "Logs"
                ],
                "summary": "Récupère les logs d'un utilisateur spécifique",
                "description": "Cette méthode permet à un administrateur de récupérer les logs d'activité d'un utilisateur spécifique basé sur son email. Elle nécessite que l'utilisateur soit authentifié et possède le rôle d'administrateur.",
                "operationId": "getLogsByUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "L'email de l'utilisateur pour lequel récupérer les logs",
                                        "type": "string",
                                        "format": "email",
                                        "example": "utilisateur@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logs récupérés avec succès pour l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "user": {
                                                "description": "Nom de l'utilisateur ayant effectué l'action",
                                                "type": "string"
                                            },
                                            "fonctionnalite": {
                                                "description": "Nom de la fonctionnalité associée à l'action",
                                                "type": "string"
                                            },
                                            "date_action": {
                                                "description": "Date et heure de l'action",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "description_action": {
                                                "description": "Description de l'action réalisée",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Utilisateur non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit pour l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Vous n'avez pas les droits pour accéder à cette ressource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des logs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Message d'erreur détaillé",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}